inference/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
ultralytics>=8.1.0
python-multipart>=0.0.9
Pillow>=10.0.0
